Implement ultra-precise 30-second smooth servo glide control
- Reconfigured servo position variables to use float for sub-microsecond precision accumulation.
- Programmed exact movement step rate of (1/30) microseconds per millisecond.
- Ensured a full 1000us sweep range (from 1000us to 2000us) takes exactly 30 seconds to complete.

+++ crsf_lora_rx/PwmController.h
@@ -35,9 +35,9 @@ private:
     uint8_t _upperSwPin;
     uint8_t _lowerSwPin;
 
-    // Smooth servo speed-limiting state variables
-    uint32_t _currentLeftUs;
-    uint32_t _currentRightUs;
+    // Smooth servo speed-limiting state variables (using floats for high precision sub-microsecond rate-limiting)
+    float _currentLeftUs;
+    float _currentRightUs;
     uint32_t _lastUpdateMs;
 };
